setup: try to catch cases where git build is not installed
This happens in cases where a new release has been tagged on master, but
the tag is not merged into the destination branch. In that case the RPM
version of the git build will be lower than what is available from the
release repos and so yum will silently skip installing the git build
that we requested.

Change-Id: I3ec30e5e5a1f01f3335b69752659ef952940609c