dogfood: prune task requirements
Yum will install all the right dependencies for us when the
beaker-integration-tests package is installed. We just need to require
python-nose (since we are invoking /usr/bin/nosetests directly) and
beaker-integration-tests itself.

The main advantage to this change is that on the new-system-page branch
where Java is no longer required, we can avoid fetching and installing
it.

Change-Id: Ib51aebfd2806052cd00976654c0912c5a5c4bc19