dogfood: run only integration tests
We already run unit tests as part of the package build, which happens
before the dogfood job is triggered. We don't need to run them again in
dogfood.

The reason to avoid running unit tests as part of the integration test
suite is that it's sometimes difficult to write unit tests which work
properly as both a regular unit test and also with the integration test
config, which is substantially different. I discovered this when working
on unit tests for bkr.server.identity in bug 1112925.

Change-Id: Iaada85c5d431e7e4c074289ba66a6314beb89061