/distribution/beaker/setup: use next major version for git builds
This will avoid spurious failures in cases where, for example, 22.2 has
been tagged but the results-reporting-improvements branch does not
contain the tag yet.

Currently the job will fail:

    :: [ 18:50:51 ] :: [ FATAL   ] :: Git build was not installed (hint: does destination branch contain latest tags?)

but now we can pass the --next-major option to Misc/rpmbuild.sh, which
was introduced during the 20.0 development cycle, and that will ensure
we always get a higher version in the git build.

Change-Id: I799e77a979e17365b0912f2ab2443e5d5b3600fa