dogfood: use __requires__ hackery for parallel installed CherryPy
This is the same hack we use on all entry points in Beaker itself as
well, with the following explanatory comment:

    # pkg_resources.requires() does not work if multiple versions are
    # installed in parallel. This semi-supported hack using __requires__
    # is the workaround.
    # http://bugs.python.org/setuptools/issue139
    # (Fedora/EPEL has python-cherrypy2 = 2.3 and python-cherrypy = 3)

The reason we have got away without needing it here in dogfood before is
because Beaker was giving us a clean RHEL6 install which didn't include
python-cherrypy (version 3).

However, on RHEL7 with EPEL enabled the python-cherrypy package now
appears in the dependency chain and we end up with both python-cherrypy2
and python-cherrypy installed, so we need this hack.

Change-Id: I68ece4a7ef21664d0e34ea6cf51fbed30781a9db