remove support for building Beaker from git
We have Jenkins to do the builds for us now. We will just assume that
whatever packages are supposed to be installed will be available from
the configured yum repos.

The sanity check to ensure that we are actually running from a git build
is valuable though, so let's keep that with a parameter to explicitly
turn it on.

Change-Id: Ic998923aab5dfface8a4b4acd63c39b73f1ea29d