dogfood: install beaker-integration-tests instead of requiring it
Currently we require beaker-integration-tests at the task level, which
means all beaker-* packages are being installed by Anaconda before any
tasks run.

But for extra realism, we should instead install the packages after
Anaconda is finished. This matches how a real Beaker server is likely to
be provisioned, and would reproduce bugs like this (which we currently
avoid in the dogfood jobs):

https://bugzilla.redhat.com/show_bug.cgi?id=1327051

Change-Id: If579df2312be048dbc7577a98480d364edb26769