Run the test as an unprivileged user
The tests in Beaker have always been executed as root. When verifying the
sandbox however, this is getting more and more out of touch with reality. The
playbooks are being run as a normal user with sudo privileges, not as root.

In order to make it more realistic, this patch creates a new user with sudo
privileges and runs the playbook under the newly created user.

Change-Id: Ifcb5d3b4b0dde453b2f278e8eab3d30caa40221b